requirements.txt
amqp>=5.0.9
asgiref>=3.5.0
async-generator>=1.10
attrs>=21.4.0
billiard>=3.6.4.0
celery>=5.2.3
certifi>=2021.10.8
cffi>=1.15.0
charset-normalizer>=2.1.0
click>=8.0.4
click-didyoumean>=0.3.0
click-plugins>=1.1.1
click-repl>=0.2.0
coreapi>=2.3.3
coreschema>=0.0.4
cryptography>=36.0.1
Deprecated>=1.2.13
Django>=4.0.6
django-admin-rangefilter>=0.9.0
django-bootstrap5>=22.2
django-crispy-forms>=1.14.0
django-debug-toolbar>=3.8.1
django-environ>=0.8.1
django-filter>=21.1
django-redis>=5.2.0
django-treebeard>=4.5.1
djangorestframework>=3.13.1
drf-yasg>=1.21.3
environ>=1.0
factory-boy>=3.2.1
Faker>=13.13.0
h11>=0.13.0
idna>=3.3
inflection>=0.5.1
itypes>=1.2.0
Jinja2>=3.1.2
kombu>=5.2.3
MarkupSafe>=2.1.1
openpyxl==3.0.9
outcome>=1.1.0
packaging>=21.3
Pillow>=9.2.0
prompt-toolkit>=3.0.28
psycopg2-binary>=2.9.3
pycparser>=2.21
pyOpenSSL>=22.0.0
pyparsing>=3.0.7
PySocks>=1.7.1
python-dateutil>=2.8.2
pytz>=2021.3
redis>=4.1.4
requests>=2.28.1
ruamel.yaml>=0.17.21
ruamel.yaml.clib>=0.2.6
selenium>=4.7.2
sentry-sdk>=1.13.0
six>=1.16.0
sniffio>=1.2.0
sortedcontainers>=2.4.0
sqlparse>=0.4.2
trio>=0.20.0
trio-websocket>=0.9.2
uritemplate>=4.1.1
urllib3>=1.26.8
vine>=5.0.0
wcwidth>=0.2.5
wrapt>=1.13.3
wsproto>=1.1.0
